Making nukleus write scheduler into its own class
